add dlp's sample data
